[Refactor] Barrel export for services
