+++ src/auth/strategies/local.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
 
-import { AuthService } from '../auth.service';
+import { AuthService } from '../services/auth.service';
 import { UserEntity } from '../../user/entities/user.entity';
 
 @Injectable()
