+++ src/user/services/user.service.spec.ts
@@ -1,10 +1,10 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UserService } from './user.service';
 import { getRepositoryToken } from '@nestjs/typeorm';
-import { UserEntity } from './entities/user.entity';
-import { UserRoleEntity } from './entities/user-role.entity';
+import { UserEntity } from '../entities/user.entity';
+import { UserRoleEntity } from '../entities/user-role.entity';
 import { ConflictException } from '@nestjs/common';
-import { PostEntity } from '../post/entities/post.entity';
+// import { PostEntity } from '../../post/entities/post.entity';
 
 describe('UserService', () => {
   let service: UserService;
